# Select dihedrals faster