… testsuite/MDAnalysisTests/core/test_atomgroup.py