Fixes (part of) #2671

Changes made in this Pull Request:
 - made phi/psi_selection longer, but faster with manual boolean array matching of atom names
 - made Ramachandran really long and ugly, but really faster with manual boolean array matching to atom names
 - no longer hardcodes phi/psi selection names in Ramachandran

The original ``Ramachandran.__init__`` with the GRO/XTC files was [180 s on my laptop.](https://github.com/MDAnalysis/mdanalysis/issues/2671#issuecomment-636708041) When I changed only the `phi_selection` code (not `psi_selection`), it dropped to 122 s. When I changed the Ramachandran code, it dropped to 0.25 s. Most of that is selecting the protein.

![Screenshot 2020-06-01 at 9 23 10 PM](https://user-images.githubusercontent.com/31115101/83404539-2dbd3380-a44e-11ea-9eb6-7a94047d5948.png)

~How much testing does this need?~ lots



PR Checklist
------------
 - [x] Tests?
 - [x] Docs?
 - [x] CHANGELOG updated?
 - [x] Issue raised/referenced?

To do
-----
- [x] update omega_selection and chi1_selection too
- [x] add `c_name`, `n_name` etc. as kwargs to those, it's annoying that the names are hard-coded in (edit: will *remove* pending *dis*approval from another dev)
- [x] add the group selection methods to `ResidueGroup`
- [x] update `Ramachandran.__init__` to use the group selection methods

Edit: Ramachandran is no longer in the slowest 50 tests.
```
========================== slowest 50 test durations ===========================

39.76s call     MDAnalysisTests/analysis/test_density.py::Test_density_from_Universe::test_notwithin

31.18s call     MDAnalysisTests/analysis/test_density.py::TestNotWithin::test_within

30.83s call     MDAnalysisTests/analysis/test_density.py::TestNotWithin::test_not_within

29.70s call     MDAnalysisTests/analysis/test_density.py::TestNotWithin::test_has_DeprecationWarning

29.58s call     MDAnalysisTests/analysis/test_hbonds.py::TestHydrogenBondAnalysis::test_true_traj

29.21s call     MDAnalysisTests/analysis/test_hbonds.py::TestHydrogenBondAnalysisPBC::test_true_traj

29.15s call     MDAnalysisTests/analysis/test_hbonds.py::TestHydrogenBondAnalysisHeuristic::test_true_traj

29.11s call     MDAnalysisTests/analysis/test_hbonds.py::TestHydrogenBondAnalysisHeavy::test_true_traj

29.11s call     MDAnalysisTests/analysis/test_hbonds.py::TestHydrogenBondAnalysisHeavyFail::test_true_traj

21.78s call     MDAnalysisTests/visualization/test_streamlines.py::test_streamplot_2D

14.44s call     MDAnalysisTests/analysis/test_hole2.py::TestHoleAnalysis::test_context_manager

12.24s call     MDAnalysisTests/coordinates/test_pdb.py::test_conect_bonds_all

10.71s call     MDAnalysisTests/analysis/test_density.py::Test_density_from_Universe::test_update_selection

9.67s call     MDAnalysisTests/analysis/test_density.py::TestDensityAnalysis::test_run[dynamic]

8.56s call     MDAnalysisTests/coordinates/test_gro.py::TestGROLargeWriter::test_writer_large

8.44s call     MDAnalysisTests/analysis/test_gnm.py::test_closeContactGNMAnalysis

8.34s call     MDAnalysisTests/analysis/test_gnm.py::test_closeContactGNMAnalysis_weights_None

7.49s call     MDAnalysisTests/analysis/test_encore.py::TestEncore::test_ces_error_estimation_ensemble_bootstrap
```