@richardjgowers 2020-06-01
This looks cool. Would it make sense to put the new logic as a ResidueGroup
method? (Psi_selections)

On Mon, Jun 1, 2020 at 12:21, Lily Wang <notifications@github.com> wrote:

> Fixes (part of) #2671
> <https://github.com/MDAnalysis/mdanalysis/issues/2671>
>
> Changes made in this Pull Request:
>
>    - made phi/psi_selection longer, but faster with manual boolean array
>    matching of atom names
>    - made Ramachandran really long and ugly, but really faster with
>    manual boolean array matching to atom names
>    - no longer hardcodes phi/psi selection names in Ramachandran
>
> The original Ramachandran.__init__ with the GRO/XTC files was 180 s on my
> laptop. When I changed only the phi_selection code, it dropped to 122 s.
> When I changed the Ramachandran code, it dropped to 0.25 s.
> [image: Screenshot 2020-06-01 at 9 15 50 PM]
> <https://user-images.githubusercontent.com/31115101/83404225-9952d100-a44d-11ea-9d75-bc42ae972f76.png>
>
> How much testing does this need?
> PR Checklist
>
>    - Tests?
>    - Docs?
>    - CHANGELOG updated?
>    - Issue raised/referenced?
>
> ------------------------------
> You can view, comment on, or merge this pull request online at:
>
>   https://github.com/MDAnalysis/mdanalysis/pull/2706
> Commit Summary
>
>    - select dihedrals faster
>
> File Changes
>
>    - *M* package/MDAnalysis/analysis/dihedrals.py
>    <https://github.com/MDAnalysis/mdanalysis/pull/2706/files#diff-ad86082633486f3093d80091122b3dc9>
>    (111)
>    - *M* package/MDAnalysis/core/topologyattrs.py
>    <https://github.com/MDAnalysis/mdanalysis/pull/2706/files#diff-96c7f631815d065c6f087f4aea01a4e6>
>    (123)
>
> Patch Links:
>
>    - https://github.com/MDAnalysis/mdanalysis/pull/2706.patch
>    - https://github.com/MDAnalysis/mdanalysis/pull/2706.diff
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/MDAnalysis/mdanalysis/pull/2706>, or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACGSGB5YOEYUQGI7BVCVYSDRUOFMZANCNFSM4NPXRAXA>
> .
>