@lilyminium 2020-06-01
@richardjgowers that's a great idea, thanks! While we're looking at `phi/psi_selection`, though, and given the recent discussion on residx vs resids, *should* we be using `segid` and `resid` to select atoms? (I kept this logic in the manual matching for now). I can see the importance of `resid` when the protein is missing a chunk in the middle, but how should `segid` be treated? If multiple segments are called the same thing, are they supposed to be treated as the same segment?