@richardjgowers 2020-06-01
I’m not enough of a bio person to really answer. I do know that resindices
make no promises on ordering, it’s just the unique number we give to each
residue.

On Mon, Jun 1, 2020 at 12:45, Lily Wang <notifications@github.com> wrote:

> @richardjgowers <https://github.com/richardjgowers> that's a great idea,
> thanks! While we're looking at phi/psi_selection, though, and given the
> recent discussion on residx vs resids, *should* we be using segid and
> resid to select atoms? (I kept this logic in the manual matching for
> now). I can see the importance of resid when the protein is missing a
> chunk in the middle, but how should segid be treated? If multiple
> segments are called the same thing, are they supposed to be treated as the
> same segment?
>
> —
> You are receiving this because you were mentioned.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/MDAnalysis/mdanalysis/pull/2706#issuecomment-636809238>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ACGSGBYYNWLB5RHBUMNZPKTRUOIEHANCNFSM4NPXRAXA>
> .
>