@lilyminium 2020-06-01
`Ramachandran.__init__` is now ~25 lines shorter but takes 0.32 s on my laptop :-( I'll let other people decide if it's worth the slowdown.